Arduino lib: resolve digitalWrite()

+++ emuino/Arduino/hardware/tools/avr/avr/include/avr/sfr_defs.h
@@ -179,7 +179,7 @@
 #define _SFR_IO8(io_addr) _MMIO_BYTE((io_addr) + __SFR_OFFSET)
 #define _SFR_IO16(io_addr) _MMIO_WORD((io_addr) + __SFR_OFFSET)
 
-#define _SFR_MEM_ADDR(sfr) ((uint16_t) &(sfr))
+#define _SFR_MEM_ADDR(sfr) ( &(sfr))
 #define _SFR_IO_ADDR(sfr) (_SFR_MEM_ADDR(sfr) - __SFR_OFFSET)
 #define _SFR_IO_REG_P(sfr) (_SFR_MEM_ADDR(sfr) < 0x40 + __SFR_OFFSET)
